Implicit flow is too implicitly described
It took me two weeks to have a working implicit OpenID Connect flow. My main errors were elements not sufficiently described (well, to my mind). I complained a lot at the lack of documentation, and would prefer not to fall in that trap again. So i tried to improve documentation on that very subject.